requirements.txt
alembic==1.12.0
bcrypt==4.0.1
blinker==1.6.2
cffi==1.16.0
click==8.1.7
cryptography==41.0.4
DateTime==5.2
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mysql==0.0.3
mysql-connector-python==8.1.0
mysqlclient==2.1.1
packaging==23.2
protobuf==4.21.12
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
SQLAlchemy==2.0.21
typing_extensions==4.8.0
Werkzeug==3.0.0
zope.interface==6.1
